core.sh (recaptcha_ack, recaptcha_nack): bash 'behavior' fix
Consider this:

f() {
  if [ -n "$1" -a "$1" -ne 0 ]; then
    echo "yes"
  fi
}

Calling this with no argument will lead to error: integer expression expected
It means that the both tests are evaluated, contrary to C language
(first test false + AND logical operand => abort, second test is not evaluated).

So I'm moving to:

f() {
  if [[ "$1" -ne 0 ]]; then
    echo "yes"
  fi
}

If argument is non-integer, it will be considered as 0.