upload.sh: skip filenames with unsupported characters
Filenames that contain commas or semicolons can't be uploaded as form
data, because curl uses those characters to separate filenames and type
information.

A workaround for commas was suggested in Issue 550 (pass data on stdin,
append a semicolon to the --form argument), but it relies on an
undocumented quirk in curl's implementation.