letitbit (download): rewrite, site obfuscates free download form
see also issue 910

Leitibit changed their site to hamper parsing the data necessary
for anonymous/free downloads. Those are obfuscated with Javascript
now, so a JS interpreter is required.

Additionally, some essential information is even hidden in a PNG
image. In order to access these data, the PNG image must be decoded
so the raw pixel color values are available. Currently, the module
can use either 'convert' (ImageMagick) [1] or 'pngtopnm' (Netpbm) [2]
for this purpose -- special thanks go to SimpleGrampgoblin for
pointing out these tools. Suggestion of alternative programs or a
solution that work directly with Bash are very welcome!

I already experimented with 'png.js' [3] which would allow this
conversion to be done by the JS interpreter. However, the library
relies on 'zlib.js' which is incompatible with the usual JS
interpreters because it uses typed array [4].

[1] http://www.imagemagick.org/www/convert.html
[2] http://netpbm.sourceforge.net/
[3] https://github.com/devongovett/png.js
[4] https://developer.mozilla.org/en-US/docs/JavaScript/Typed_arrays/